Adding CLI Based Orchestration (#36)
This change introduces a state machine that will help the user step through their deployment. Right now, it is a loop of plan, confirm, deploy. 

The LLM is used to create a plan, we analyze the project and then use the LLM to present a summary. A user will then be able to confirm the deployment and a deployment will trigger. Which now just prints to the screen.

It also introduces go-workflows for executing actions.