Supabase Functions for Returning Docker Tokens (#45)
This change introduces two supabase edge functions that our CLI can use to retrieve tokens that can be used for pushing and pulling from a users private docker repo that we manage.

The role setup on the AWS side will be extracted to some IaC and we'll update the functions when we tie into supabase auth.