Slight Workflow Refactor (#58)
The initial way I wrote the deploy workflow relied on closures and dynamically registering and executing activities. This can be non-deterministic and can cause some challenges with go-workflows, and in generally a bad pattern.

Rewrote it to wrap https://github.com/meroxa/prod/blob/main/cli/internal/deployment/render/executor.go#L30 in a single activity and that activity is registered at startup like the other activities.

The other change is that I moved the non-error console output into activities instead of in the workflow. The go-workflow has a replay functionality that will replay a workflow (without running the activities again) and this will cause double writing to the console.