making sure DB and webservice are in the same region. Adding status checking to ensure DB is ready (#78)
In testing a full stack app deployment, found we needed to make a couple of small updates:

Make sure the web service and the DB are deployed in the same region. If they are not, then the web service will not be able to reach the DB over the private connection string.
Add status checking to the deployment of the DB. This makes sure it is ready before the web service tries to connect.