Feat: Switch UI to Use Bubble Tea. (#81)
This PR introduces bubbletea. It refactors the interface to be a richer TUI.

There are a few main changes:

the tui/ package contains the functionality for rendering. model.go does the majority of the heavily lifting and right now contains both the model and view code. The view can be split out later as we evolve how things look.
The unified writer interface. It is a bit hacky right now but works.