Environment Variable Detection (#106)
This PR is all about environment variables. It introduces the ability to detect them in the target base and use the LLM to categorize them. Once categorized, we can smartly set in connection strings, etc from the PaaS providers.

For non database related environment variables, we prompt the user to enter the values, using what they have in .env files as defaults.