Cleanup TUI Table (#108)
The first iteration used lots of ASCII to make the table. This cleans
that up to use lipgloss to align the columns and rows.

I tried using the table component from bubbletea but it was giving
all kinds of trouble, so went with this "handrolled" approach, since are
needs are quite simple right now.