Integrating Auth Into Interactive CLI Flow (#112)
Integrates: #110 into the CLI flow.

There are a few key changes:

Agent.Process will check for validate authentication on every input.
On valid authentication, the Session struct will be injected into the context. We need to use a workflow.Propagator to make sure it travels between normal go contexts and workflow contexts
The workflows and activities can access the session from the context
The backend client APIs now take an access token for making calls to the edge functions
The edge functions will check for a valid JWT
The record-stack edge function will use the authenticated role instead of the system role
Still to do:
Make the creation of the ECR repos for each each user/project based on the current user, instead of hardcoded user