Bug fixes (#123)
Apologies for the big PR :)

This PR contains several bug fixes from some testing of deploying more projects:

Local images for our base docker images
If a user has the image we use for the base image on their machine, our docker build will try to use it. This is an issue if the architecture of their local image doesn't match amd64, which we need for Render deploys. We'll end up with an error about mismatched architectures Instead of deleting their local image or force pulling amd64, I've created our own tagged images we can host that will be amd64. So there is no/low chance of mismatching architecture.
Timeout Categorizing Environment Variables
When there were a number of environment variables we ran the risk of the workflow timing out as we processed all the env vars in one action. To fix this, I increase the timeout overall, as well as run the categorization in parallel.
Accurate endpoint checking
After we deploy to a PaaS, we will check to see if the app is live. What I was doing was using the url provided back by render or fly. This was always the root. However, if we are deploying things that don't have the / as a route (like an API), this will always end up with Prod reporting a failure. We now do some smart LLM based categorization of the routes to get a best match on what we should check. We then use that path to poll.
TUI display of pricing
Small fix, but moved the displaying of the pricing data in our TUI tables.