Python Docker Deploys (#142)
This PR adds the ability to have python projects deployed by Docker, tested using our Render support.

Key Changes:

Determining the run command. This is done in two phases. First we'll do static analysis on the project for common approaches for launching, like main . Next, once we have determined the candidates we will pass it to the LLM with some additional context like the README contents and environment variables. The LLM will then give us back what it thinks we should use as the run command to launch the project.
'Robust' dependency management tool support. There a few different dependency management tools in play with python. Need to handle this in the dockerfile, so we do some additional static analysis before rendering the template.
Updated route detection for python projects. When a router is being used the paths in the files can be relative, so we need to apply further analysis to find the actual full paths.