Support For Web Projects that Have a Build Step (#143)
This PR introduces the ability to deploy web projects that have a build step. Projects that use typescript and/or bundlers should deploy successfully.

Tested with docker builds to Render with the following:

Vue
Svelte
Astro
Angular

I've introduced some more LLM analysis that will try to determine where the build output lands. We can then use this further to determine if the app can be deployed as a static app, and in that case we'll use an nginx base image to run things. Otherwise, we can drop back to our node based image when we know there is a server involved.