Using slog for structured logging (#159)
We were using a mix of log.Print/log.Println/log.Printf statements for logging. This PR converts all those to using structured logs (default slog).

As a nice to have, I also added an embedded log viewer that is enabled in debug mode to make searching on our side a bit easier as we develop.